--- src/pythonfinder/models/asdf.py
@@ -1,9 +1,0 @@
-# -*- coding=utf-8 -*-
-import attr
-
-from .pyenv import PyenvFinder
-
-
-@attr.s
-class AsdfFinder(PyenvFinder):
-    version_root = attr.ib(default="installs/python/*")
