+++ src/pythonfinder/utils.py
@@ -7,10 +7,12 @@ import os
 from fnmatch import fnmatch
 
 import attr
+import io
 import six
 
 import vistir
 
+from .environment import PYENV_INSTALLED, PYENV_ROOT, ASDF_INSTALLED, ASDF_DATA_DIR
 from .exceptions import InvalidPythonVersion
 
 try:
@@ -127,12 +129,6 @@ def filter_pythons(path):
     return filter(lambda x: path_is_python(x), path.iterdir())
 
 
-# def unnest(item):
-#     if isinstance(next((i for i in item), None), (list, tuple)):
-#         return chain(*filter(None, item))
-#     return chain(filter(None, item))
-
-
 def unnest(item):
     if isinstance(item, Iterable) and not isinstance(item, six.string_types):
         item, target = itertools.tee(item, 2)
@@ -145,3 +141,30 @@ def unnest(item):
                 yield sub
         else:
             yield el
+
+
+def parse_pyenv_version_order(filename="version"):
+    version_order_file = normalize_path(os.path.join(PYENV_ROOT, filename))
+    if os.path.exists(version_order_file) and os.path.isfile(version_order_file):
+        with io.open(version_order_file, encoding="utf-8") as fh:
+            contents = fh.read()
+        version_order = [v for v in contents.splitlines()]
+        return version_order
+
+
+def parse_asdf_version_order(filename=".tool-versions"):
+    version_order_file = normalize_path(os.path.join("~", filename))
+    if os.path.exists(version_order_file) and os.path.isfile(version_order_file):
+        with io.open(version_order_file, encoding="utf-8") as fh:
+            contents = fh.read()
+        python_section = next(iter(
+            line for line in contents.splitlines() if line.startswith("python")
+        ), None)
+        if python_section:
+            python_key, versions = python_section.partition()
+            if versions:
+                return versions.split()
+
+
+def is_in_path(path, parent):
+    return normalize_path(str(path)).startswith(normalize_path(str(parent)))
